requirements.txt
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
matplotlib==3.10.3
numpy==2.2.5
plotly==6.1.0
streamlit==1.45.1
streamlit-chat==0.1.1
streamlit-javascript==0.1.5
toml==0.10.2
python-dotenv==1.1.1